+++ Api/src/main/java/com/example/springboot/repositories/CategoriaRepository.java
@@ -3,8 +3,6 @@ package com.example.springboot.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.springboot.models.CategoriaModel;
 
-
-
-public interface CategoriaRepository extends JpaRepository<CategoriaModel, Integer> {
+public interface CategoriaRepository extends JpaRepository<CategoriaModel, Long>{
     
 }
